requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.21
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pandas==2.1.1
openpyxl==3.1.2
python-dotenv==1.0.0
email-validator==2.0.0
alembic==1.12.0
pytest==7.4.2
httpx==0.25.0
pytest-cov==4.1.0
